Revert "Monsters"
This reverts commit 3c820e58067acd3ef2888316e9439c050a7e5361.

+++ src/entities/Bloc.java
@@ -111,8 +111,8 @@ public abstract class Bloc extends Entity{
 		}
 	}
 
-	public double[] getCollisionArea(){
-		double[] res = new double[4];
+	public int[] getCollisionArea(){
+		int[] res = new int[4];
 		int x;
 		if (width >= length)  x = width;
 		else x = length;
